Upload the basic communication system
I tried to implement the communication function in the uploaded files. But here are some problems. The AI kept saying nonsense (probably because of the poor prompt). I did not use the muti-thread environment, which I will implement in the following code. Since the basic functions have been implemented. I will try the system you instructed.