Added the demo of the alignment
After trying to manage the whole repository, I gave up.   XD.....(we will start a new repository anyway right?)    This is the demo I write to implement the  AI alignment system. Flask is the API that runs locally to call other API. There are some test files in the System, plz ignore them. I applied the Maven structure, so you can check the dependencies in the Pom file. I put a test data set in AlignmentSystem. But it is far from satisfying TAT. But it at least implements the basic function that the agent could interact with each other.